requirements.txt
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.9
django-cors-headers==2.4.0
django-heroku==0.3.1
django-otp==0.5.0
django-templated-mail==1.1.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djoser==1.3.0
gunicorn==19.9.0
idna==2.7
oauthlib==2.1.0
psycopg2==2.7.7
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
rest-social-auth==1.4.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.23
whitenoise==4.1.2
